+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Develop]
+
+### Added
+
+- added `customIdFunction` as parameter ([#2])
+- added example for `customIdFunction` in README.md
+
+### Fixed
+
+- link for 1.1.0 compare in CHANGELOG.md
+
 ## [1.1.0] (2020-11-03)
 
 ### Added
@@ -29,4 +40,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - added 'formatter'
 - added tests for <b>only one</b> of the three branches from the formatter
 
+[#2]: https://github.com/LetsMelon/addIdToArray/issues/2
+
+[develop]: https://github.com/LetsMelon/addIdToArray/compare/main...develop
 [1.1.0]: https://github.com/LetsMelon/addIdToArray/compare/v1.0.0...v1.1.0